Code now is clean
